♻️ refactor : 파일 생성해서 관리
