+++ src/styles/components/header_style/SearchInputStyle.jsx
