+++ src/styles/components/header_style/UserProfileStyle.jsx
